gateway_requirements_minimal.txt
# API Gateway Requirements - Minimal for testing
# Lightweight gateway with minimal dependencies

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.3,<3.0.0
httpx>=0.24.0  # For service communication

# PDF processing - minimal
pdf2image==1.16.3
Pillow>=10.0.0  # Image processing